more test consolidation
